# Conflicts:
#	plugins/Institution/src/Model/Table/InstitutionShiftsTable.php
#	plugins/Institution/src/Model/Table/StaffAttendancesTable.php
#	plugins/Institution/src/Model/Table/StudentAttendancesTable.php

+++ plugins/Security/src/Controller/Component/SelectOptionsTamperingComponent.php
@@ -28,7 +28,7 @@ class SelectOptionsTamperingComponent extends Component {
                     $session->delete('FormTampering');
                     $session->write('FormTamperingReload', $requestData);
                 }
-            } else {
+            } else if (!$this->request->is('ajax')) {
                 $session->delete('FormTampering');
                 $session->delete('FormTamperingReload');
             }
